refactor(cua-computer): split browser action files
